refactor(native): move the interactive tools into their own module
Permission action mapping, `ask_user_question`, `request_plan_exit`, their two
input validators, `ApprovalOutcome`, `await_approval`, and `plan_mode_denial`
move to `api_process_adapter/interactive.rs`.

Measured at 326 lines, not the 370 design.md estimated: that range ran 37 lines
past `plan_mode_denial` into `parse_optional_non_negative_integer_arg` and
`non_negative_integer`, which parse tool-call arguments and go to
`native_tools.rs` instead.
